refactor: auth 모듈을 db / dep / user로 분리
