+++ backend/src/auth/dependency.py
@@ -1,29 +1,10 @@
-from typing import Annotated, Any, AsyncGenerator
+from typing import Annotated
 
-import fastapi_users
 from fastapi import Depends
-from fastapi_users import BaseUserManager
-from fastapi_users.authentication import JWTStrategy
-from fastapi_users.db.base import BaseUserDatabase
-from fastapi_users_db_sqlmodel import SQLModelUserDatabaseAsync
 
-from ..config import settings
-from ..dependency import SessionDep
-from .model import OAuthAccount, User
-from .user import UserManager
+from .model import User
+from .user import fastapi_users
 
+current_active_verified_user = fastapi_users.current_user(active=True, verified=True)
 
-async def get_user_db(session: SessionDep) -> AsyncGenerator[SQLModelUserDatabaseAsync[User, int], Any]:
-    yield SQLModelUserDatabaseAsync(session, User, OAuthAccount)
-
-
-async def get_user_manager(account_db: "UserDBDep") -> AsyncGenerator[UserManager, Any]:
-    yield UserManager(account_db)
-
-
-def get_jwt_strategy() -> JWTStrategy:
-    return JWTStrategy(secret=settings.JWT_SECRET, lifetime_seconds=3600)
-
-
-UserDBDep = Annotated[BaseUserDatabase, Depends(get_user_db)]
-UserManagerDep = Annotated[BaseUserManager, Depends(get_user_manager)]
+CURRENT_USER = Annotated[User, Depends(current_active_verified_user)]
